scripts/requirements.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
db-dtypes==1.4.3
fonttools==4.60.1
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.33.1
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.3
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
statsmodels==0.14.5
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
